+++ components/Canvas.tsx
@@ -7,13 +7,12 @@ import React, {
   useRef,
   useState,
 } from "react";
-import styles from "./Canvas.module.scss";
 export interface CanvasProps {}
 
 var isDrawing : boolean = false;
 var lastX : number | null;
 var lastY : number | null;
-var paths : Array<any> = [];
+var paths : Array<Array<any>> = [];
 
 const getDrawConfig : any = (e: MouseEvent, brush: IBrush) => {
   return {
@@ -34,8 +33,7 @@ const getDrawConfig : any = (e: MouseEvent, brush: IBrush) => {
 const draw = (ctx: CanvasRenderingContext2D, cfg: any, canvas: HTMLCanvasElement) => {
   makePath(ctx, cfg);
   
-  lastX = cfg.x;
-  lastY = cfg.y;
+  
   switch(cfg.effect){
     case 'mirror-x' : return mirrorX(ctx, cfg, canvas); break;
     case 'mirror-y' : return mirrorY(ctx, cfg, canvas); break;
@@ -43,30 +41,35 @@ const draw = (ctx: CanvasRenderingContext2D, cfg: any, canvas: HTMLCanvasElement
   }
 }
 
-const makePath = (ctx: CanvasRenderingContext2D, cfg: any,) => {
+const makePath = (ctx: CanvasRenderingContext2D, cfg: any) => {
   ctx.globalCompositeOperation = cfg.operation;
   ctx.globalAlpha = cfg.alpha;
   ctx.fillStyle = cfg.color;
   ctx.strokeStyle = cfg.color;
   ctx.lineWidth = cfg.width;
-  ctx.lineCap = cfg.shape;
+  ctx.lineCap = cfg.shape === 'round' ? 'round' : 'square';
+  ctx.lineJoin = cfg.shape === 'round' ? 'round' : 'bevel';
+  ctx.imageSmoothingQuality = 'high';
   ctx.beginPath();
-
-  if(cfg.shape === 'square'){
-    ctx.rect(cfg.x - cfg.width / 2, cfg.y - cfg.width / 2, cfg.width, cfg.height);
+  if (lastX && lastY && (cfg.x !== lastX || cfg.y !== lastY)) {
+    ctx.moveTo(lastX, lastY);
+    ctx.lineTo(cfg.x, cfg.y);
+    ctx.stroke();
   }
-  else if(cfg.shape === 'round'){
-    ctx.arc(cfg.x - cfg.width / 16, cfg.y - cfg.width / 16, cfg.width /2, 0, Math.PI * 2);
+  else{
+    if(cfg.shape === 'square'){
+      ctx.rect(cfg.x - cfg.width / 2, cfg.y - cfg.width / 2, cfg.width, cfg.height);
+      console.log(cfg.x - cfg.width / 2, cfg.y - cfg.width / 2, cfg.width, cfg.height)
+    }
+    else{
+      ctx.arc(cfg.x, cfg.y, cfg.width /2, 0, Math.PI * 2);
+    }
+    ctx.fill();
   }
-  // connect pints when mouse moves fast
-  // if (lastX && lastY && (cfg.x !== lastX || cfg.y !== lastY)) {
-  //   ctx.moveTo(lastX, lastY);
-  //   ctx.lineTo(cfg.x, cfg.y);
-  // }
   ctx.closePath();
-  ctx.fill();
-  paths.push(cfg);
-  // ctx.stroke(); 
+  lastX = cfg.x;
+  lastY = cfg.y;
+  paths[paths.length-1].push(cfg);
 }
 
 const mirrorX = (ctx: CanvasRenderingContext2D, cfg: any, canvas: HTMLCanvasElement) => {
@@ -98,16 +101,20 @@ const clearCanvas = (ctx: CanvasRenderingContext2D, canvas: HTMLCanvasElement) =
   paths = [];
 }
 
-
 const undoPath = (ctx: CanvasRenderingContext2D, canvas: HTMLCanvasElement) => {
-  let refresh_paths = paths.slice(0, -1);
+  let new_paths = paths.slice(0, -1);
   clearCanvas(ctx, canvas);
-  refresh_paths.forEach(cfg => makePath(ctx, cfg));
-}
-
-const fitCanvas = (canvas: HTMLCanvasElement) => {
-  canvas.height = window.innerHeight;
-  canvas.width = window.innerWidth - 220;
+  if(new_paths.length > 0){
+    paths[0] = [];
+    new_paths.forEach((new_path, i) => {
+      new_path.forEach(cfg => makePath(ctx, cfg));
+      lastX = null;
+      lastY = null;
+      if(i < new_paths.length - 1){
+        paths.push([]);
+      }
+    });
+  }
 }
 
 const saveImageWithName = (canvas: HTMLCanvasElement) => {
@@ -121,23 +128,23 @@ const saveImageWithName = (canvas: HTMLCanvasElement) => {
   }
 }
 
+
 export const Canvas = ({}: CanvasProps) => {
   const canvasRef = useRef<HTMLCanvasElement | null>(null);
+  const cursorRef = useRef<HTMLDivElement | null>(null);
   const [ctx, setContext] = useState<CanvasRenderingContext2D | null>(null);
   const [brush, setBrush] = useState<IBrush>(INIT_BRUSH);
   const { data, setData } = useContext(DataContext);
   const [link, setLink] = useState<string>("");
   const downRef = useRef<{ x: number; y: number } | null>(null);
-
   useEffect(() => {
-    // init canvas dimensions and drawing context
     const canvas = canvasRef.current;
     if (canvas) {
       const ctx = canvas.getContext("2d") as CanvasRenderingContext2D;
       setContext(ctx);
-      fitCanvas(canvas);
       canvas.addEventListener("mousedown", e => {
         isDrawing = true;
+        paths.push([])
         draw(ctx, getDrawConfig(e, brush), canvas);
       });
       canvas.addEventListener("mousemove", e => {
@@ -145,11 +152,15 @@ export const Canvas = ({}: CanvasProps) => {
           draw(ctx, getDrawConfig(e, brush), canvas);
         }
       });
-      canvas.addEventListener("mouseup", () => {
-        isDrawing = false;
+      canvas.addEventListener("mouseup", () =>  {
+        isDrawing = false, 
+        lastX = null;
+        lastY = null;
       });
-      canvas.addEventListener("mouseleave", () => {
-        isDrawing = false;
+      canvas.addEventListener("mouseleave", () =>  {
+        isDrawing = false, 
+        lastX = null;
+        lastY = null;
       });
       canvas.addEventListener("contextmenu", e => e.preventDefault());
       window.addEventListener("keydown", e => {
@@ -158,19 +169,15 @@ export const Canvas = ({}: CanvasProps) => {
         }
       })
       window.addEventListener('clear-canvas', () => clearCanvas(ctx, canvas));
-      window.addEventListener('resize', () => fitCanvas(canvas));
       window.addEventListener('save-as', () => saveImageWithName(canvas));
     }
   }, []);
 
   return (
-    <div className={styles.root}>
-      <div className={styles.container}>
-        <canvas
-          className={styles.container}
-          ref={canvasRef}
-        />
-      </div>
-    </div>
+    <canvas className='canvas-layer'
+      ref={canvasRef}
+      height='600'
+      width='800'
+    />
   );
 };
