--- components/canvas/Canvas.module.scss
@@ -1,15 +1,0 @@
-.root {
-  background-color: #000000;
-}
-
-.header {
-  display: flex;
-}
-
-
-.container {
-  background-color: #ffffff;
-  user-select: none;
-  display: flex;
-}
-
