--- components/header/Header.module.scss
@@ -1,32 +1,0 @@
-.root {
-  border-bottom: 1px solid #ddd;
-  background-color: rgba(0, 0, 0, 0.05);
-  padding: 4px;
-  display: flex;
-  gap: 4px;
-}
-
-.container {
-  padding: 0.5rem;
-}
-
-.dropdown{
-  background-color: red;
-}
-
-.dropdownContent{
-  background-color: lightgrey;
-  border-radius: 2px;
-  padding: 4px;
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  border: 1px solid darkgrey;
-  margin-top: 2px;
-  // display: none;
-  button{
-    border: none;
-    background-color: none;
-  }
-}
