+++ components/index.ts
@@ -1,4 +1,4 @@
-export * from "./canvas/Canvas";
-export * from "./header/Header";
-export * from "./layout/Layout";
-export * from "./tools/Tools";
+export * from "./Canvas";
+export * from "./TopBar";
+export * from "./Layout";
+export * from "./Tools";
