--- components/tools/Tools.module.scss
@@ -1,67 +1,0 @@
-.root {
-  border-left: 1px solid #ddd;
-  background-color: rgba(0, 0, 0, 0.05);
-  width: 220px;
-}
-
-.container {
-  display: grid;
-  row-gap: 0.5rem;
-  padding: 0.5rem;
-}
-
-.colors {
-  display: flex;
-  align-items: center;
-}
-
-.field {
-  display: grid;
-}
-
-.select {
-  border: 1px solid #ddd;
-}
-
-.label {
-  display: block;
-  padding: 0.25rem;
-  display: flex;
-  font-size: 0.75rem;
-  font-weight: 600;
-  align-items: center;
-}
-
-.input {
-  width: 100%;
-  &[type="color"] {
-    background-color: #ddd;
-    border: 1px solid #ddd;
-    cursor: pointer;
-    height: 2.5rem;
-    overflow: hidden;
-    width: 2.5rem;
-    &::-webkit-color-swatch {
-      border: unset;
-    }
-    &::-webkit-color-swatch-wrapper {
-      padding: 0;
-    }
-  }
-}
-
-.window {
-  background-color: rgba(0, 0, 0, 0.05);
-  border: 1px solid #ddd;
-  display: grid;
-  place-items: center;
-  height: calc(220px - 1rem);
-  width: calc(220px - 1rem);
-}
-
-.brush {
-  background-color: rgba(255, 255, 255, 0.5);
-
-  border: 1px solid #000;
-  border-style: dashed;
-}
