+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import "@styles/globals.scss";
+import "@styles/_globals.scss";
 import type { AppProps } from "next/app";
 
 function MyApp({ Component, pageProps }: AppProps) {
