--- styles/Home.module.scss
@@ -1,5 +1,0 @@
-.root {
-}
-
-.container {
-}
