+++ styles/Tools.scss
@@ -1,0 +1,71 @@
+#tools {
+  background-color: lightgrey;
+  width: 220px;
+  padding: 4px;
+  height: 100vh;
+  position: absolute;
+  right: 0px;
+  top: 0px;
+  overflow: auto;
+  label {
+    display: block;
+    display: flex;
+    font-size: 0.75rem;
+    font-weight: 600;
+    align-items: center;
+  }
+  input, select, button {
+    width: 100%;
+    &[type="color"] {
+      background-color: #ddd;
+      border: 1px solid #ddd;
+      cursor: pointer;
+      height: 2.5rem;
+      overflow: hidden;
+      width: 2.5rem;
+      &::-webkit-color-swatch {
+        border: unset;
+      }
+      &::-webkit-color-swatch-wrapper {
+        padding: 0;
+      }
+    }
+  }
+  .window {
+    background-color: rgba(0, 0, 0, 0.05);
+    border: 1px solid #ddd;
+    display: grid;
+    place-items: center;
+    height: calc(220px - 1rem);
+    width: calc(220px - 1rem);
+  }
+  .brush {
+    background-color: rgba(255, 255, 255, 0.5);
+  
+    border: 1px solid #000;
+    border-style: dashed;
+  }
+  .container {
+    display: grid;
+    row-gap: 0.5rem;
+  }
+  
+  .colors {
+    display: flex;
+    align-items: center;
+  }
+  
+  .field {
+    display: grid;
+  }
+  
+  .select {
+    border: 1px solid #ddd;
+  }
+
+  .tools-section{
+    .section-label{
+      font-weight: bold;
+    }
+  }
+}
