+++ styles/TopBar.scss
@@ -1,0 +1,38 @@
+#top-bar{
+    font-size: 12px;
+    background-color: lightgrey;
+    padding: 4px;
+    button {
+        font-size: 12px;
+        border: none;
+        outline: none;
+        padding: 2px;
+        border-radius: 2px;
+        &:hover{
+            color: white;
+            background-color:rgba(0, 0, 0, 0.35);
+        }
+    }
+    #top-bar-menu{
+        display: flex;
+        gap: 4px;
+        .dropdown-content{
+            background-color: lightgrey;
+            border-radius: 2px;
+            padding: 4px;
+            position: absolute;
+            display: flex;
+            flex-direction: column;
+            gap: 4px;
+            border: none;
+            border-radius: 2px;
+            margin-top: 2px;
+            -webkit-box-shadow: 0px 7px 4px 4px rgba(161,161,161,0.6); 
+            box-shadow: 0px 7px 4px 4px rgba(161,161,161,0.6);
+            button{
+                border: none;
+                background-color: none;
+            }
+        }
+    }
+}
